requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
Jinja2==3.1.2
secure-smtplib==0.1.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
